Branch merge.
Everything seems to have gone smoothly so far. Had to move setting the listener ref up in the Event Beginplay Sequence in BP_Character so it properly sets for the fist character but a quick test run looks good. Lemme know if anything broke under the hood that I didn't notice.